+++ projects/spotfire-wrapper/src/public_api.ts
@@ -1,5 +1,5 @@
 /*
-* Copyright (c) 2019. TIBCO Software Inc.
+* Copyright (c) 2019-2020. TIBCO Software Inc.
 * This file is subject to the license terms contained
 * in the license file that is distributed with this file.
 */
@@ -13,5 +13,5 @@ export * from './lib/editor/spotfire-editor.component';
 export * from './lib/viewer/spotfire-viewer.module';
 export * from './lib/editor/spotfire-editor.module';
 export { DocumentService } from './lib/document.service';
-export { SpotfireParameters, SpotfireReporting } from './lib/spotfire-webplayer';
+export { SpotfireParameters, SpotfireReporting, SpotfireFiltering } from './lib/spotfire-webplayer';
 export { SpotfireCustomization } from './lib/spotfire-customization';